grub.cfg: Move memtest to utilities, fix condition
memtest is located in the "Utilities" menu when booted in BIOS mode (ie.
syslinux bootloader), however when booted in UEFI mode (ie. GRUB
bootloader), it's in the "Advanced Installation" menu, which is
surprising.

So move it to Utilities, and also fix the condition to show it (it's not
only for BIOS systems, it's also for UEFI). Which basically get us close
to the file shipped by live-build itself.